requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
Faker
